master

+++ travis/test_server.py
@@ -2,6 +2,7 @@
 
 from __future__ import print_function
 
+import ast
 import re
 import os
 import shutil
@@ -82,14 +83,14 @@ def has_test_errors(fname, dbname, odoo_version, check_loaded=True):
                 ignore = True
                 break
         if ignore:
-            break
+            continue
         for report_pattern in errors_report:
             if report_pattern(log_record):
                 errors.append(log_record)
                 break
 
     if check_loaded:
-        if not [r for r in log_records if 'Modules loaded.' == r['message']]:
+        if not [r for r in log_records if 'Modules loaded.' in r['message']]:
             errors.append({'message': "Modules loaded message not found."})
 
     if errors:
@@ -129,9 +130,9 @@ def get_addons_path(travis_dependencies_dir, travis_build_dir, server_path):
     :param server_path: Server path
     :return: Addons path
     """
-    addons_path_list = get_addons(travis_dependencies_dir)
-    addons_path_list.insert(0, travis_build_dir)
-    addons_path_list.append(server_path + "/addons")
+    addons_path_list = get_addons(travis_build_dir)
+    addons_path_list.extend(get_addons(travis_dependencies_dir))
+    addons_path_list.append(os.path.join(server_path, "addons"))
     addons_path = ','.join(addons_path_list)
     return addons_path
 
@@ -176,11 +177,32 @@ def get_test_dependencies(addons_path, addons_list):
                 os.path.join(path, addons_list[0]))
             if not manif_path:
                 continue
-            manif = eval(open(manif_path).read())
+            manif = ast.literal_eval(open(manif_path).read())
             return list(
                 set(manif.get('depends', [])) |
                 set(get_test_dependencies(addons_path, addons_list[1:])) -
                 set(addons_list))
+
+
+def cmd_strip_secret(cmd):
+    cmd_secret = []
+    skip_next = False
+    for param in cmd:
+        if skip_next:
+            skip_next = False
+            continue
+        if param.startswith('--db_'):
+            cmd_secret.append(param.split('=')[0] + '=***')
+            continue
+        if param.startswith('--log-db'):
+            cmd_secret.append('--log-db=***')
+            continue
+        if param in ['-w', '-r']:
+            cmd_secret.extend([param, '***'])
+            skip_next = True
+            continue
+        cmd_secret.append(param)
+    return cmd_secret
 
 
 def setup_server(db, odoo_unittest, tested_addons, server_path, script_name,
@@ -216,7 +238,7 @@ def setup_server(db, odoo_unittest, tested_addons, server_path, script_name,
                      "--stop-after-init",
                      "--init", ','.join(preinstall_modules),
                      ] + install_options + server_options
-        print(" ".join(cmd_odoo))
+        print(" ".join(cmd_strip_secret(cmd_odoo)))
         subprocess.check_call(cmd_odoo)
     return 0
 
@@ -282,6 +304,8 @@ def main(argv=None):
     unbuffer = str2bool(os.environ.get('UNBUFFER', True))
     data_dir = os.environ.get("DATA_DIR", '~/data_dir')
     test_enable = str2bool(os.environ.get('TEST_ENABLE', True))
+    dbtemplate = os.environ.get('MQT_TEMPLATE_DB', 'openerp_template')
+    database = os.environ.get('MQT_TEST_DB', 'openerp_test')
     if not odoo_version:
         # For backward compatibility, take version from parameter
         # if it's not globally set
@@ -324,22 +348,21 @@ def main(argv=None):
     else:
         print("Modules to test: %s" % tested_addons)
     # setup the base module without running the tests
-    dbtemplate = "openerp_template"
     preinstall_modules = get_test_dependencies(addons_path,
                                                tested_addons_list)
-    preinstall_modules = list(set(preinstall_modules) - set(get_modules(
-        os.environ.get('TRAVIS_BUILD_DIR'))))
+
+    preinstall_modules = list(set(preinstall_modules or []) - set(get_modules(
+        os.environ.get('TRAVIS_BUILD_DIR')) or [])) or ['base']
     print("Modules to preinstall: %s" % preinstall_modules)
     setup_server(dbtemplate, odoo_unittest, tested_addons, server_path,
                  script_name, addons_path, install_options, preinstall_modules,
                  unbuffer, server_options)
 
     # Running tests
-    database = "openerp_test"
-
     cmd_odoo_test = ["coverage", "run",
                      "%s/%s" % (server_path, script_name),
                      "-d", database,
+                     "--db-filter=^%s$" % database,
                      "--stop-after-init",
                      "--log-level", test_loglevel,
                      ]
@@ -386,12 +409,12 @@ def main(argv=None):
                 command_call = [item
                                 for item in commands[0][0]
                                 if item not in rm_items] + \
-                    ['--db-filter=^%s$' % database]
+                    ['--pidfile=/tmp/odoo.pid']
             else:
                 command[-1] = to_test
                 # Run test command; unbuffer keeps output colors
                 command_call = (["unbuffer"] if unbuffer else []) + command
-            print(' '.join(command_call))
+            print(" ".join(cmd_strip_secret(command_call)))
             pipe = subprocess.Popen(command_call,
                                     stderr=subprocess.STDOUT,
                                     stdout=subprocess.PIPE)
@@ -415,7 +438,7 @@ def main(argv=None):
                 counted_errors += errors
                 all_errors.append(to_test)
                 print(fail_msg, "Found %d lines with errors" % errors)
-        if not instance_alive:
+        if not instance_alive and odoo_unittest:
             # Don't drop the database if will be used later.
             subprocess.call(["dropdb", database])
 
@@ -433,6 +456,7 @@ def main(argv=None):
         return 1
     # if we get here, all is OK
     return 0
+
 
 if __name__ == '__main__':
     exit(main())
